# backtiendavirtual/swagger.yml
openapi: 3.0.3
info:
  title: Tienda Virtual API
  description: |
    Documentación de la API para la tienda virtual.
    Incluye endpoints para autenticación, gestión de productos, usuarios, carrito de compras y más.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Servidor local por defecto
...
paths:
...
  /api/products:
    get:
      tags:
        - Products
      summary: Obtiene todos los productos (operación pública)
      operationId: getAllProducts
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
    post:
      tags:
        - Products
      summary: Crea un producto (ejemplo de endpoint público, si así lo configuras)
      operationId: createProductPublic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Producto creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorsResponse'
        '409':
          description: Conflicto (producto ya existe)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
...
  /api/adminTools/products:
    get:
      tags:
        - AdminTools
      summary: Obtiene la lista de productos paginada (solo ADMIN)
      operationId: getProductsPaginate
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número de página
          required: false
          schema:
            type: integer
            default: 1
        - name: name
          in: query
          description: Filtro por nombre de producto (máx. 50 caracteres)
          required: false
          schema:
            type: string
            maxLength: 50
      responses:
        '200':
          description: Lista de productos paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: object
                    properties:
                      totalPages:
                        type: integer
                      actualPage:
                        type: integer
                  products:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Error de validación de parámetros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorsResponse'
        '401':
          description: Usuario bloqueado o no autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /api/adminTools/users:
    get:
      tags:
        - AdminTools
      summary: Obtiene la lista de usuarios paginada (solo ADMIN)
      operationId: getUsersPaginate
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Número de página (1 por defecto)
          required: false
          schema:
            type: integer
            default: 1
        - name: username
          in: query
          description: Filtro por nombre de usuario (máx. 50 caracteres)
          required: false
          schema:
            type: string
            maxLength: 50
      responses:
        '200':
          description: Lista de usuarios paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  pages:
                    type: object
                    properties:
                      totalPages:
                        type: integer
                      actualPage:
                        type: integer
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminToolsTypeUserResponse'
        '400':
          description: Error de validación en parámetros
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorsResponse'
        '401':
          description: Usuario bloqueado o no autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '500':
          description: Error interno del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /api/adminTools/users/{username}:
    delete:
      tags:
        - AdminTools
      summary: Bloquea o desbloquea un usuario (solo ADMIN)
      operationId: updateUserBlocked
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: Nombre de usuario que se desea bloquear/desbloquear
          schema:
            type: string
      responses:
        '200':
          description: Operación exitosa (Usuario bloqueado o desbloqueado)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '401':
          description: Usuario bloqueado o no autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /api/adminTools/product/{id}:
    delete:
      tags:
        - AdminTools
      summary: Bloquea o desbloquea un producto (solo ADMIN)
      operationId: blockProductAdmin
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID del producto a bloquear/desbloquear
          schema:
            type: integer
      responses:
        '200':
          description: Producto bloqueado o desbloqueado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '404':
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '401':
          description: No autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
  /api/adminTools/users/update/bussiness/{username}:
    put:
      tags:
        - AdminTools
      summary: Añade o quita el rol de negocio a un usuario (solo ADMIN)
      operationId: updateUserToBussiness
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: path
          required: true
          description: Username al que se le agregará o quitará el rol
          schema:
            type: string
      responses:
        '200':
          description: Actualización correcta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '401':
          description: Usuario bloqueado o no autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '404':
          description: Usuario no encontrado o rol no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '500':
          description: Error interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'


  ###########################################################################
  #                                LABELS
  ###########################################################################
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    GenericResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: boolean
    ValidationErrorsResponse:
      allOf:
        - $ref: '#/components/schemas/GenericResponse'
        - type: object
          properties:
            errors:
              type: object
              additionalProperties:
                type: string
          example:
            message: "Errores de validación en la solicitud"
            status: 400
            error: true
            errors:
              fieldName: "Mensaje de error concreto"

    ####################################################
    #        RESPUESTAS / MODELOS ESPECÍFICOS
    ####################################################
...
    AdminToolsTypeUserResponse:
      type: object
      properties:
        username:
          type: string
        isBussiness:
          type: boolean
        isAdmin:
          type: boolean
        isBlocked:
          type: boolean
...
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        image:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
        sold:
          type: integer
        productOwner:
          type: string
        categories:
          type: array
          items:
            $ref: '#/components/schemas/CategorySimpleResponse'
        labels:
          type: array
          items:
            $ref: '#/components/schemas/LabelSimpleResponse'
        productBlocked:
          type: boolean
          description: Indica si el producto se encuentra bloqueado
        userOwnerBlocked:
          type: boolean
          description: Indica si el dueño de ese producto se encuentra bloqueado
...
    CategorySimpleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
    LabelSimpleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string

    ####################################################
    #   ENTIDADES BÁSICAS (PUEDEN SER DE SOLO LECTURA)
    ####################################################
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        image:
          type: string
        price:
          type: number
          format: double
        quantity:
          type: integer
        sold:
          type: integer
        blocked:
          type: boolean
...